requirements.txt
# requirements.txt
anthropic>=0.42.0
streamlit==1.28.1
python-dotenv==1.0.0
requests==2.31.0